+++ ichnaea/api/locate/tests/test_fallback.py
@@ -1,22 +1,288 @@
+import colander
 import mock
 import requests_mock
-import simplejson as json
 from redis import RedisError
 from requests.exceptions import RequestException
 
 from ichnaea.api.exceptions import LocationNotFound
 from ichnaea.api.locate.constants import DataSource
-from ichnaea.api.locate.fallback import FallbackPositionSource
+from ichnaea.api.locate.fallback import (
+    DisabledCache,
+    ExternalResult,
+    FallbackCache,
+    FallbackPositionSource,
+    OutboundSchema,
+    ResultSchema,
+)
+from ichnaea.api.locate.query import Query
 from ichnaea.api.locate.result import Position
 from ichnaea.api.locate.tests.base import (
     BaseSourceTest,
     DummyModel,
 )
+from ichnaea.api.locate.tests.test_query import QueryTest
+from ichnaea import floatjson
+from ichnaea.tests.base import TestCase
 from ichnaea.tests.factories import (
     ApiKeyFactory,
     CellFactory,
     WifiFactory,
 )
+
+
+class TestExternalResult(TestCase):
+
+    def test_not_found(self):
+        result = ExternalResult(None, None, None, None)
+        self.assertTrue(result.not_found())
+
+    def test_not_found_accuracy(self):
+        result = ExternalResult(1.0, 1.0, None, None)
+        self.assertTrue(result.not_found())
+
+    def test_found(self):
+        result = ExternalResult(1.0, 1.0, 10, None)
+        self.assertFalse(result.not_found())
+
+    def test_found_fallback(self):
+        result = ExternalResult(1.0, 1.0, 10, 'lacf')
+        self.assertFalse(result.not_found())
+
+
+class TestResultSchema(TestCase):
+
+    def setUp(self):
+        super(TestResultSchema, self).setUp()
+        self.schema = ResultSchema()
+
+    def test_empty(self):
+        with self.assertRaises(colander.Invalid):
+            self.schema.deserialize({})
+
+    def test_accuracy_float(self):
+        data = self.schema.deserialize(
+            {'location': {'lat': 1.0, 'lng': 1.0}, 'accuracy': 11.6})
+        self.assertEqual(
+            data, {'lat': 1.0, 'lon': 1.0, 'accuracy': 11, 'fallback': None})
+
+    def test_accuracy_missing(self):
+        with self.assertRaises(colander.Invalid):
+            self.schema.deserialize(
+                {'location': {'lat': 1.0, 'lng': 1.0}, 'fallback': 'lacf'})
+
+    def test_fallback(self):
+        data = self.schema.deserialize(
+            {'location': {'lat': 1.0, 'lng': 1.0},
+             'accuracy': 10, 'fallback': 'lacf'})
+        self.assertEqual(
+            data, {'lat': 1.0, 'lon': 1.0, 'accuracy': 10, 'fallback': 'lacf'})
+
+    def test_fallback_invalid(self):
+        data = self.schema.deserialize(
+            {'location': {'lat': 1.0, 'lng': 1.0},
+             'accuracy': 10, 'fallback': 'cidf'})
+        self.assertEqual(
+            data, {'lat': 1.0, 'lon': 1.0, 'accuracy': 10, 'fallback': None})
+
+    def test_fallback_missing(self):
+        data = self.schema.deserialize(
+            {'location': {'lat': 1.0, 'lng': 1.0}, 'accuracy': 10})
+        self.assertEqual(
+            data, {'lat': 1.0, 'lon': 1.0, 'accuracy': 10, 'fallback': None})
+
+    def test_location_incomplete(self):
+        with self.assertRaises(colander.Invalid):
+            self.schema.deserialize(
+                {'location': {'lng': 1.0}, 'accuracy': 10, 'fallback': 'lacf'})
+
+    def test_location_missing(self):
+        with self.assertRaises(colander.Invalid):
+            self.schema.deserialize({'accuracy': 10, 'fallback': 'lacf'})
+
+
+class TestOutboundSchema(TestCase):
+
+    def setUp(self):
+        super(TestOutboundSchema, self).setUp()
+        self.schema = OutboundSchema()
+
+    def test_empty(self):
+        self.assertEqual(self.schema.deserialize({}), {})
+        self.assertEqual(self.schema.deserialize({'unknown_field': 1}), {})
+
+    def test_fallback(self):
+        self.assertEqual(self.schema.deserialize(
+            {'fallbacks': {'ipf': False}}),
+            {'fallbacks': {}})
+        self.assertEqual(self.schema.deserialize(
+            {'fallbacks': {'lacf': False}}),
+            {'fallbacks': {'lacf': False}})
+        self.assertEqual(self.schema.deserialize(
+            {'fallbacks': {'ipf': True, 'lacf': False}}),
+            {'fallbacks': {'lacf': False}})
+
+    def test_query(self):
+        query = Query()
+        data = self.schema.deserialize(query.internal_query())
+        self.assertEqual(data, {'fallbacks': {'lacf': True}})
+
+    def test_cell(self):
+        cell = CellFactory.build()
+        query = Query(cell=[
+            {'radio': cell.radio, 'mcc': cell.mcc, 'mnc': cell.mnc,
+             'lac': cell.lac, 'cid': cell.cid, 'signal': -70, 'ta': 15,
+             'unknown_field': 'foo'}])
+        data = self.schema.deserialize(query.internal_query())
+        self.assertEqual(data, {
+            'cellTowers': [{
+                'radioType': cell.radio.name,
+                'mobileCountryCode': cell.mcc,
+                'mobileNetworkCode': cell.mnc,
+                'locationAreaCode': cell.lac,
+                'cellId': cell.cid,
+                'signalStrength': -70,
+                'timingAdvance': 15,
+            }],
+            'fallbacks': {'lacf': True},
+        })
+
+    def test_wifi(self):
+        wifis = WifiFactory.build_batch(2)
+        query = Query(wifi=[
+            {'key': wifi.key, 'signal': -90} for wifi in wifis])
+        data = self.schema.deserialize(query.internal_query())
+        self.assertEqual(data, {
+            'wifiAccessPoints': [{
+                'macAddress': wifis[0].key,
+                'signalStrength': -90,
+            }, {
+                'macAddress': wifis[1].key,
+                'signalStrength': -90,
+            }],
+            'fallbacks': {'lacf': True},
+        })
+
+
+class TestCache(QueryTest):
+
+    def setUp(self):
+        super(TestCache, self).setUp()
+        self.cache = FallbackCache(
+            self.raven_client, self.redis_client, self.stats_client,
+            cache_expire=600)
+
+    def test_disabled(self):
+        cache = DisabledCache()
+        query = Query()
+        self.assertEqual(
+            cache.set(query, ExternalResult(None, None, None, None)), None)
+        self.assertEqual(cache.get(query), None)
+
+    def test_get_cell(self):
+        cells = CellFactory.build_batch(1)
+        query = Query(cell=self.cell_model_query(cells))
+        self.assertEqual(self.cache.get(query), None)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:miss']),
+        ])
+
+    def test_set_cell(self):
+        cell = CellFactory.build()
+        query = Query(cell=self.cell_model_query([cell]))
+        result = ExternalResult(cell.lat, cell.lon, cell.range, None)
+        self.cache.set(query, result)
+        keys = self.redis_client.keys('cache:fallback:cell:*')
+        self.assertEqual(len(keys), 1)
+        self.assertTrue(500 < self.redis_client.ttl(keys[0]) <= 600)
+        self.assertEqual(self.cache.get(query), result)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:hit']),
+        ])
+
+    def test_set_cell_not_found(self):
+        cell = CellFactory.build()
+        query = Query(cell=self.cell_model_query([cell]))
+        result = ExternalResult(None, None, None, None)
+        self.cache.set(query, result)
+        keys = self.redis_client.keys('cache:fallback:cell:*')
+        self.assertEqual(len(keys), 1)
+        self.assertEqual(self.redis_client.get(keys[0]), b'"404"')
+        self.assertEqual(self.cache.get(query), result)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:hit']),
+        ])
+
+    def test_get_cell_multi(self):
+        cells = CellFactory.build_batch(2)
+        query = Query(cell=self.cell_model_query(cells))
+        self.assertEqual(self.cache.get(query), None)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:bypassed']),
+        ])
+
+    def test_get_wifi(self):
+        wifis = WifiFactory.build_batch(2)
+        query = Query(wifi=self.wifi_model_query(wifis))
+        self.assertEqual(self.cache.get(query), None)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:miss']),
+        ])
+
+    def test_set_wifi(self):
+        wifis = WifiFactory.build_batch(2)
+        wifi = wifis[0]
+        query = Query(wifi=self.wifi_model_query(wifis))
+        result = ExternalResult(wifi.lat, wifi.lon, wifi.range, None)
+        self.cache.set(query, result)
+        self.assertEqual(self.cache.get(query), result)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:hit']),
+        ])
+
+    def test_set_wifi_inconsistent(self):
+        wifis1 = WifiFactory.build_batch(2)
+        self.cache.set(
+            Query(wifi=self.wifi_model_query(wifis1)),
+            ExternalResult(wifis1[0].lat, wifis1[0].lon, 100, None))
+
+        # similar lat/lon, worse accuracy
+        wifis2 = WifiFactory.build_batch(
+            2, lat=wifis1[0].lat + 0.0001, lon=wifis1[0].lon)
+        self.cache.set(
+            Query(wifi=self.wifi_model_query(wifis2)),
+            ExternalResult(wifis2[0].lat, wifis2[0].lon, 200, None))
+
+        # check combined query, avg lat/lon, max accuracy
+        query = Query(wifi=self.wifi_model_query(wifis1 + wifis2))
+        self.assertEqual(
+            self.cache.get(query),
+            ((wifis1[0].lat + wifis2[0].lat) / 2, wifis1[0].lon, 200, None))
+
+        # different lat/lon
+        wifis3 = WifiFactory.build_batch(2, lat=wifis1[0].lat + 10.0)
+        self.cache.set(
+            Query(wifi=self.wifi_model_query(wifis3)),
+            ExternalResult(wifis3[0].lat, wifis3[0].lon, 300, None))
+
+        # check combined query, inconsistent result
+        query = Query(wifi=self.wifi_model_query(wifis1 + wifis2 + wifis3))
+        self.assertEqual(self.cache.get(query), None)
+
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:hit']),
+            ('locate.fallback.cache', 1, 1, ['status:inconsistent']),
+        ])
+
+    def test_get_mixed(self):
+        cells = CellFactory.build_batch(1)
+        wifis = WifiFactory.build_batch(2)
+        query = Query(
+            cell=self.cell_model_query(cells),
+            wifi=self.wifi_model_query(wifis))
+        self.assertEqual(self.cache.get(query), None)
+        self.check_stats(counter=[
+            ('locate.fallback.cache', 1, 1, ['status:bypassed']),
+        ])
 
 
 class TestSource(BaseSourceTest):
@@ -46,15 +312,45 @@ class TestSource(BaseSourceTest):
             'accuracy': self.fallback_model.range,
             'fallback': 'lacf',
         }
+        self.fallback_cached_result = floatjson.float_dumps({
+            'lat': self.fallback_model.lat,
+            'lon': self.fallback_model.lon,
+            'accuracy': self.fallback_model.range,
+            'fallback': 'lacf',
+        })
 
     def _mock_redis_client(self):
         client = mock.Mock()
-        client.pipeline.return_value = mock.Mock()
-        client.pipeline.return_value.__enter__ = mock.Mock()
-        client.pipeline.return_value.__exit__ = mock.Mock()
+        client.pipeline.return_value = client
+        client.__enter__ = mock.Mock(return_value=client)
+        client.__exit__ = mock.Mock(return_value=None)
+        client.expire.return_value = mock.Mock()
         client.get.return_value = mock.Mock()
+        client.mget.return_value = mock.Mock()
         client.set.return_value = mock.Mock()
+        client.mset.return_value = mock.Mock()
         return client
+
+    def test_cache(self):
+        source = self.TestSource(
+            settings={'cache_expire': '60'},
+            geoip_db=self.geoip_db,
+            raven_client=self.raven_client,
+            redis_client=self.redis_client,
+            stats_client=self.stats_client,
+        )
+        self.assertTrue(isinstance(source.cache, FallbackCache))
+        self.assertEqual(source.cache.cache_expire, 60)
+
+    def test_no_cache(self):
+        source = self.TestSource(
+            settings={'cache_expire': '0'},
+            geoip_db=self.geoip_db,
+            raven_client=self.raven_client,
+            redis_client=self.redis_client,
+            stats_client=self.stats_client,
+        )
+        self.assertTrue(isinstance(source.cache, DisabledCache))
 
     def test_success(self):
         cell = CellFactory.build()
@@ -248,7 +544,7 @@ class TestSource(BaseSourceTest):
             mock_request.register_uri(
                 'POST', requests_mock.ANY, json=self.fallback_result)
 
-            ratelimit_key = self.source._get_ratelimit_key()
+            ratelimit_key = self.source._ratelimit_key()
             self.redis_client.set(ratelimit_key, self.source.ratelimit)
 
             query = self.model_query(cells=[cell])
@@ -276,40 +572,50 @@ class TestSource(BaseSourceTest):
     def test_get_cache_redis_failure(self):
         cell = CellFactory.build()
         mock_redis_client = self._mock_redis_client()
-        mock_redis_client.get.side_effect = RedisError()
+        mock_redis_client.mget.side_effect = RedisError()
 
         with requests_mock.Mocker() as mock_request:
             mock_request.register_uri(
                 'POST', requests_mock.ANY, json=self.fallback_result)
 
-            with mock.patch.object(self.source, 'redis_client',
+            with mock.patch.object(self.source.cache, 'redis_client',
                                    mock_redis_client):
                 query = self.model_query(cells=[cell])
                 result = self.source.search(query)
                 self.check_model_result(result, self.fallback_model)
 
-            self.assertTrue(mock_redis_client.get.called)
+            self.assertTrue(mock_redis_client.mget.called)
             self.assertTrue(mock_request.called)
+
+        self.check_stats(counter=[
+            ('locate.fallback.cache', ['status:failure']),
+        ])
 
     def test_set_cache_redis_failure(self):
         cell = CellFactory.build()
         mock_redis_client = self._mock_redis_client()
-        mock_redis_client.get.return_value = None
-        mock_redis_client.set.side_effect = RedisError()
+        mock_redis_client.mget.return_value = []
+        mock_redis_client.mset.side_effect = RedisError()
+        mock_redis_client.expire.side_effect = RedisError()
+        mock_redis_client.execute.side_effect = RedisError()
 
         with requests_mock.Mocker() as mock_request:
             mock_request.register_uri(
                 'POST', requests_mock.ANY, json=self.fallback_result)
 
-            with mock.patch.object(self.source, 'redis_client',
+            with mock.patch.object(self.source.cache, 'redis_client',
                                    mock_redis_client):
                 query = self.model_query(cells=[cell])
                 result = self.source.search(query)
                 self.check_model_result(result, self.fallback_model)
 
-            self.assertTrue(mock_redis_client.get.called)
-            self.assertTrue(mock_redis_client.set.called)
+            self.assertTrue(mock_redis_client.mget.called)
+            self.assertTrue(mock_redis_client.mset.called)
             self.assertTrue(mock_request.called)
+
+        self.check_stats(counter=[
+            ('locate.fallback.cache', ['status:miss']),
+        ])
 
     def test_cache_single_cell(self):
         cell = CellFactory.build()
@@ -363,8 +669,6 @@ class TestSource(BaseSourceTest):
             self.check_stats(counter=[
                 ('locate.fallback.cache', ['status:miss']),
                 ('locate.fallback.lookup', ['status:404']),
-            ], timer=[
-                'locate.fallback.lookup',
             ])
 
             query = self.model_query(cells=[cell])
@@ -375,70 +679,26 @@ class TestSource(BaseSourceTest):
             self.check_stats(counter=[
                 ('locate.fallback.cache', ['status:hit']),
                 ('locate.fallback.lookup', ['status:404']),
-            ], timer=[
-                'locate.fallback.lookup',
             ])
 
     def test_dont_recache(self):
         cell = CellFactory.build()
         mock_redis_client = self._mock_redis_client()
-        mock_redis_client.get.return_value = json.dumps(self.fallback_result)
+        mock_redis_client.mget.return_value = [self.fallback_cached_result]
 
         with requests_mock.Mocker() as mock_request:
             mock_request.register_uri(
                 'POST', requests_mock.ANY, json=self.fallback_result)
 
-            with mock.patch.object(self.source, 'redis_client',
+            with mock.patch.object(self.source.cache, 'redis_client',
                                    mock_redis_client):
                 query = self.model_query(cells=[cell])
                 result = self.source.search(query)
                 self.check_model_result(result, self.fallback_model)
 
-            self.assertTrue(mock_redis_client.get.called)
-            self.assertFalse(mock_redis_client.set.called)
-
-    def test_cache_expire_0(self):
-        cell = CellFactory.build()
-        mock_redis_client = self._mock_redis_client()
-        mock_redis_client.get.return_value = json.dumps(self.fallback_result)
-        self.source.cache_expire = 0
-
-        with requests_mock.Mocker() as mock_request:
-            mock_request.register_uri(
-                'POST', requests_mock.ANY, json=self.fallback_result)
-
-            with mock.patch.object(self.source, 'redis_client',
-                                   mock_redis_client):
-                query = self.model_query(cells=[cell])
-                result = self.source.search(query)
-                self.check_model_result(result, self.fallback_model)
-
-            self.assertFalse(mock_redis_client.get.called)
-            self.assertFalse(mock_redis_client.set.called)
-
-    def test_cache_no_wifi(self):
-        cell = CellFactory.build()
-        wifis = WifiFactory.build_batch(2)
-        mock_redis_client = self._mock_redis_client()
-        mock_redis_client.get.return_value = json.dumps(self.fallback_result)
-
-        with requests_mock.Mocker() as mock_request:
-            mock_request.register_uri(
-                'POST', requests_mock.ANY, json=self.fallback_result)
-
-            with mock.patch.object(self.source, 'redis_client',
-                                   mock_redis_client):
-
-                query = self.model_query(cells=[cell], wifis=wifis)
-                result = self.source.search(query)
-                self.check_model_result(result, self.fallback_model)
-
-            self.assertFalse(mock_redis_client.get.called)
-            self.assertFalse(mock_redis_client.set.called)
+            self.assertTrue(mock_redis_client.mget.called)
+            self.assertFalse(mock_redis_client.mset.called)
 
         self.check_stats(counter=[
-            ('locate.fallback.cache', ['status:bypassed']),
-            ('locate.fallback.lookup', ['status:200']),
-        ], timer=[
-            'locate.fallback.lookup',
+            ('locate.fallback.cache', ['status:hit']),
         ])